Changed amount from int to float

+++ KetKoin/Transaction.cs
@@ -7,7 +7,7 @@ namespace KetKoin;
 public class Transaction : BlockData
 {
     public int TransactionNumber { get; set; }
-    public int Amount { get; set; }
+    public float Amount { get; set; }
     public DateTime TimeStamp { get; set; }
     public Byte[] SenderKey { get; set; }
     public Byte[] RecieverKey { get; set; }
@@ -15,7 +15,7 @@ public class Transaction : BlockData
     public Type Type { get; set; }
 
 
-    public Transaction(int amount, Byte[] senderKey, Byte[] senderPrivateKey, Byte[] recieverKey,int transactionNumber, Type type)
+    public Transaction(float amount, Byte[] senderKey, Byte[] senderPrivateKey, Byte[] recieverKey,int transactionNumber, Type type)
     {
         TransactionNumber = transactionNumber;
         Amount = amount;
